Change the window name to avoid window name collisions with IDEs
When having a sway config rule to move the application window to a convenient portion of the workspace the rule might try to match window names. If for instance, an IDE is opened it will ikely put the project name in its title, if only for a short while. During that time the window rule intended for the application will also match the IDEs window ans will modify its properties accordingly. Having the window name differ from the project name will reduce this risk.